docs(reference): match the exports — hooks rewritten from source, completeness guard re-armed
The reference pages drifted from packages/. Every claim below was re-verified
against source before the edit.

- cli.mdx: delete `vendo extract` (never existed) and `login --email` (not a
  flag); `--framework` gains `custom`; doctor `--json` never starts the dev
  server even with `--yes`; add `vendo knowledge` and `vendo config`, sourced
  from the CLI's own HELP.
- hooks.mdx: rewritten from packages/ui/src/hooks. No `execute` callback
  exists; `useApps` is exportApp/importApp; `useVendoStatus` returns
  {posture, connected}; `useVendoThread` is sendMessage; `data` is seeded and
  `error` is undefined-not-null; polling is per-hook and never pauses on a
  hidden tab; send-queueing lives in the chrome composer, not the hook.
- server-api.mdx / handler-options.mdx: HostOAuthAdapter rides the root entry;
  `development` is a boolean; `model` is optional and deprecated; the anon
  cookie is Path=/ on purpose; the server base path is a hard constant;
  ephemeral principals DO write to disk; /box/* is the fourth CSRF exception;
  structuredRepair and smokeRender are on by default; the composition table
  gains the seven missing keys.
- http-routes.mdx: adds heartbeat, /approvals/:id, /connections/catalog,
  fn proxy, machine ping, review queue, reject-review, the doctor probes and
  the /box callback surface; /sync/impact is 403 + {impact}, not 404.
- dot-vendo.mdx: drops the nonexistent enriched/watermark fields, adds
  judgments.json and four more real files, states the three-layer split.
- environment-variables.mdx: adds VENDO_MODEL_KNOWLEDGE_VERIFIER and the other
  host-facing knobs; drops the "every env var" completeness claim it could not
  keep.

The guard: handler-options.docs.test.ts's type assertions lived in a test file,
which tsconfig.json excludes from the program — tsc never saw them, which is
how seven public keys went undocumented under a green suite. They now live in
handler-options.docs-check.ts, which tsconfig.docs-check.json typechecks.
