+++ docs-site/reference/dot-vendo.mdx
@@ -1,25 +1,45 @@
 ---
-title: "The .vendo directory: tools, policy, theme, and data files"
+title: "The .vendo directory: tools, judgments, policy, theme, and data files"
 sidebarTitle: ".vendo directory"
-description: "Every path inside the .vendo directory, which tool writes it, and the exact format contract, from tools.json and policy.json to theme.json and data."
+description: "Every path inside the .vendo directory, which tool writes it, and the exact format contract, from tools.json and judgments.json to policy.json, theme.json, and data."
 ---
+
+`.vendo/` is split by **author**, in three layers. Knowing which layer a file
+belongs to tells you whether you may edit it and what can overwrite it:
+
+| Layer | Written by | Files |
+| --- | --- | --- |
+| Machine | the deterministic scan (`vendo sync`) — regenerated wholesale | `tools.json`, `catalog.json`, `theme.extracted.json`, `knowledge-manifest.json` |
+| Judgment | the AI pass — regenerated by the judge, never by a human | `judgments.json` |
+| Human | you (init scaffolds some of them) — never overwritten | `overrides.json`, `policy.json`, `brief.md`, `design-rules.md`, `theme.json`, `knowledge.json` |
+
+That split is the whole point: a re-sync can rewrite the machine layer and
+re-judge the AI layer without ever clobbering a decision a person made.
 
 | Path | Written by | Format |
 | --- | --- | --- |
-| `tools.json` | sync | `vendo/tools@3` (04) — tools with inferred field `semantics`, per-tool `srcHash`, a per-tool `enriched` marker (set by the AI enrichment pass; absent = structural defaults), and the `watermark` (git tree at the last enrichment; absent until one runs) |
+| `tools.json` | sync | `vendo/tools@3` (04) — each tool's descriptor and execution `binding`, plus inferred field `semantics`, `srcHash` (the handler-source hash incremental sync diffs against), `audience`, `outputSchema` when the source declares one, and `disabled` for anything extraction could not classify |
+| `judgments.json` | sync (AI pass) | `vendo/judgments@1` (04) — the AI layer, keyed by tool name: `description`, `title`, `risk`, `critical`, `disabled`, `audience`, `semantics`. Hardenings apply themselves; loosenings only land here after a human accepted them |
+| `overrides.json` | init / human | `vendo/overrides@3` (04) — per-tool overrides plus `compounds`, `briefs`, per-surface tool menus (`surfaces.agent` / `surfaces.mcp`, see [Curate the tool menu](/capabilities/mcp#curate-the-tool-menu)), and `remix.ignoreSlots` |
 | `catalog.json` | sync | `vendo/catalog@1` (04) |
-| `overrides.json` | init / human | `vendo/overrides@3` (04) — per-tool overrides plus `compounds`, `briefs`, and per-surface tool menus (`surfaces.agent` / `surfaces.mcp`, see [Curate the tool menu](/capabilities/mcp#curate-the-tool-menu)) |
 | `policy.json` | init / human | `vendo/policy@1` (05) |
 | `remixable/<slot>.json` | sync | `PinBaseline` (06 §8) |
 | `brief.md` | init | product brief for the system prompt (03 §3) |
 | `design-rules.md` | host, optional | generation-time design rules (06 §5), re-read per generation so edits apply without a restart; `createVendo({ apps: { designRules } })` wins over it |
-| `theme.json` | init extraction | `VendoTheme` |
+| `theme.json` | init extraction / human | `VendoTheme` |
+| `theme.extracted.json` | init / sync | `vendo/theme-extracted@1` — what the deterministic scan produced last time. It is the merge base that lets sync tell "you edited this slot" from "the app's brand moved"; commit it beside `theme.json` |
+| `knowledge.json` | `vendo knowledge add` / human | the registered knowledge sources (globs, kind, visibility, name) |
+| `knowledge-manifest.json` | `vendo knowledge sync` | sha256 hashes of every synced document. Regenerable — safe to gitignore; deleting it re-syncs everything |
 | `data/` | store | PGlite files (gitignored; init writes the ignore) |
 | `data/misses.jsonl` | runtime | one `vendo/capability-miss@1` event per line ([Telemetry](/deploy/telemetry#capability-misses)) |
 
-This table is complete, and it is split by author: `tools.json` is the machine
-layer `vendo sync` regenerates wholesale, and `overrides.json` is the only
-hand-edited file, so extraction can never overwrite reviewed risk,
-descriptions, disabled state, compounds, or briefs.
-
 `data/` is runtime state, not configuration. Do not commit it.
+
+Five of these files are also **cloud-resolvable** surfaces:
+`design-rules.md`, `brief.md`, `theme.json`, `policy.json`, and
+`overrides.json`. Per surface, resolution is an explicit `createVendo` value →
+the local file → the console's published value → unset, so the file's
+existence is the switch. `vendo config status` shows which layer owns each
+one; see [the CLI reference](/reference/cli#vendo-config-command-dir).
+`tools.json` and `catalog.json` are deliberately not on that list — they are
+generation inputs, not host-editable content.
