+++ docs-site/reference/environment-variables.mdx
@@ -1,8 +1,12 @@
 ---
 title: "Environment variables for Vendo host, sandbox, and peers"
 sidebarTitle: "Environment variables"
-description: "Every environment variable Vendo reads across host process, app machines, and peer providers, with the effect each has on the composition and CLI."
+description: "The environment variables a host developer sets across host process, app machines, and peer providers, with the effect each has on the composition and CLI."
 ---
+
+These are the variables a host developer sets. Vendo also reads a handful of
+internal test seams and harness markers that are not listed here and are not
+part of any contract — if a name is not on this page, do not depend on it.
 
 ## Host process
 
@@ -14,7 +18,7 @@ description: "Every environment variable Vendo reads across host process, app ma
 | `VENDO_TICK_SECRET` | bearer secret required by `POST /tick` |
 | `VENDO_BASE_URL` | trusted host origin used for server-side execution of extracted route tools, for marking the anonymous-session cookie `Secure` behind a TLS-terminating proxy, as the MCP door's canonical public base, and as the origin of the box callback URLs (`VENDO_STORE_URL`/`VENDO_HOST_URL`); machine provisioning fails without it |
 | `VENDO_URL` | base URL that `vendo doctor` probes for the live `/status` check and that `vendo sync` queries for blast-radius impact; defaults to `http://localhost:3000/api/vendo`, overridden per run by `--url` |
-| `VENDO_POSTHOG_KEY` | overrides the telemetry project key |
+| `VENDO_POSTHOG_KEY` · `VENDO_POSTHOG_HOST` | override the telemetry project key and the host capture events are sent to |
 | `DO_NOT_TRACK=1` | disables telemetry and is also honored by Scarf |
 | `CI` | disables telemetry |
 | `NODE_ENV=production` | disables build and development telemetry at runtime |
@@ -28,7 +32,11 @@ description: "Every environment variable Vendo reads across host process, app ma
 | `VERCEL`, `CF_PAGES`, `AWS_LAMBDA_FUNCTION_NAME` | prevent the default PGlite store from opening on an ephemeral serverless filesystem; pass a Postgres URL instead |
 | `VENDO_STORE_ENCRYPTION_KEY` | 32-byte base64 key the default composition uses to encrypt secret values in `vendo_secrets`; read when no store is passed to `createVendo`. An explicitly configured store ignores it |
 | `VENDO_DEV_CREDENTIAL` | pins a [dev-mode credential ladder](/connect/dev-mode) rung: `env-key:anthropic` (or `:openai` / `:google`), `vendo-cloud`, or `none` |
-| `VENDO_MODEL` · `VENDO_MODEL_PAINT` · `VENDO_MODEL_JUDGE` · `VENDO_MODEL_EXTRACT` | per-slot model pins, no code change: the string passes verbatim to whatever the resolved credential talks to (the Vendo Cloud gateway with `VENDO_API_KEY`, your provider otherwise). Precedence per slot: explicit model object → env pin → `models` string → per-rung default. `VENDO_MODEL_EXTRACT` also pins the `vendo init` / `vendo sync` extraction engine |
+| `VENDO_MODEL` · `VENDO_MODEL_PAINT` · `VENDO_MODEL_JUDGE` · `VENDO_MODEL_KNOWLEDGE_VERIFIER` · `VENDO_MODEL_EXTRACT` | per-slot model pins, no code change: the string passes verbatim to whatever the resolved credential talks to (the Vendo Cloud gateway with `VENDO_API_KEY`, your provider otherwise). Precedence per slot: explicit model object → env pin → `models` string → per-rung default. `VENDO_MODEL_EXTRACT` also pins the `vendo init` / `vendo sync` extraction engine; `VENDO_EXTRACTION_MODEL` is its older spelling and still works |
+| `VENDO_KNOWLEDGE_VERIFY` | `on` runs the knowledge-search verifier pass, which costs a model call and seconds of latency per search; **off by default**. Any value other than on/true/1 or off/false/0 throws at compose rather than silently meaning off. The verifier resolves on the `knowledgeVerifier` model slot above; a slot that resolves to nothing simply marks results unverified |
+| `VENDO_AWAY_TOKEN_SECRET` | default source for the away-token preset's shared secret — the one the producer mints with and the host's verifier checks. Pass `secret` explicitly to override it; see [actAs presets](/connect/act-as-presets) |
+| `VENDO_TURN_IDLE_ABORT_MS` | how long a streaming turn may go without a client heartbeat before the server aborts it (default 15,000). Raise it on a runtime where heartbeats are unreliable; the watchdog only ever arms after a first heartbeat |
+| `VENDO_APP_BUILD_WATCHDOG_MS` | how long one app build may run before it is failed with a terminal, retryable record (default 4 minutes). The client's polling cutoff derives from it, so raising this keeps the two in step |
 | `VENDO_CLOUD_MODEL` | **deprecated** (use `VENDO_MODEL`) — agent-slot pin for the Vendo Cloud gateway; the gateway remaps unknown ids during the alias grace window |
 | `ANTHROPIC_API_KEY` · `OPENAI_API_KEY` · `GOOGLE_GENERATIVE_AI_API_KEY` | first-preferred rung of the credential ladder read by `vendoModel()`; when set, `@vendoai/vendo/server` uses the matching `@ai-sdk/*@^3` provider installed in your app for the full native tool loop |
 
@@ -36,6 +44,13 @@ description: "Every environment variable Vendo reads across host process, app ma
 `VENDO_BOX_*` knobs are implementation-level environment configuration. The
 frozen contracts define the corresponding bearer, route origin, and telemetry
 seams without requiring those names from custom compositions.
+
+The `auth` presets read their provider's own environment rather than a
+Vendo-named variable — `AUTH_SECRET` for Auth.js, `CLERK_SECRET_KEY` for
+Clerk, `SUPABASE_JWT_SECRET` / `SUPABASE_URL` for Supabase,
+`AUTH0_DOMAIN` (or `AUTH0_ISSUER_BASE_URL`) for Auth0. `jwt()` has no vendor
+env to read and always takes an explicit secret. See
+[actAs presets](/connect/act-as-presets) for the full table.
 
 ## App machine environment
