+++ docs-site/reference/server-api.mdx
@@ -15,10 +15,13 @@ import type { ActionsRegistry, Connector } from "@vendoai/actions";
 import type { VendoGuard, PolicyConfig, Judge } from "@vendoai/guard";
 import type { AppsRuntime, SandboxAdapter } from "@vendoai/apps";
 import type { AutomationsEngine } from "@vendoai/automations";
-import type { HostAuthPreset, HostOAuthAdapter } from "@vendoai/vendo/server";
+// HostOAuthAdapter and ToolRegistry ride the ROOT entry, not /server.
+import type { HostOAuthAdapter, ToolRegistry } from "@vendoai/vendo";
+import type { ConnectionsService, HostAuthPreset, ModelsConfig, ServerActionHandler } from "@vendoai/vendo/server";
 
 export function createVendo(config: {
-  model: LanguageModel;                       // the one required thing
+  models?: ModelsConfig;                      // { agent?, paint?, judge?, knowledgeVerifier? } — a model name or an AI SDK model object
+  model?: LanguageModel;                      // deprecated alias of models.agent; with neither, the credential ladder resolves one from the environment
   auth?: HostAuthPreset;                      // one preset fills principal + actAs + oauth; see /connect/act-as-presets
   principal?: (req: Request) => Promise<Principal | null>;   // escape hatch, mutually exclusive with auth; null → ephemeral anonymous
   store?: VendoStore;                         // default: createStore() (PGlite, .vendo/data)
@@ -34,15 +37,17 @@ export function createVendo(config: {
   oauth?: HostOAuthAdapter;                   // escape hatch, mutually exclusive with auth; required when mcp is true and auth supplies no oauth half
   serverActions?: Record<string, ServerActionHandler>; // generated by `vendo sync`
   agent?: {                                   // chat context controls; see Handler options
+    instructions?: string;                    // host voice, appended to the system prompt every turn
     toolOutputCap?: number;                   // default 32000; 0 disables the cap
     maxOutputTokens?: number;
     historyWindow?: number;                   // last N whole messages re-sent per turn
     maxInitialTools?: number;                 // initial tool loadout cap; rest via vendo_tools_search
+    maxSearchExpansions?: number;             // lazy toolkits one tool search may expand (default 3)
     maxSteps?: number;                        // agent-loop step cap per turn (default 20)
   };
   sessions?: { ttlMs?: number; sweepIntervalMs?: number }; // anonymous-session lifecycle
   approvals?: { parkedCallTtlMs?: number };   // BYO parked-call expiry; default 60 min, 0 disables
-  development?: boolean | { root?: string; out?: string }; // dev-only source capture
+  development?: boolean;                      // dev-only injection seams; NODE_ENV=development enables them
 }): Vendo;
 
 export interface Vendo {
@@ -54,10 +59,13 @@ export interface Vendo {
 ```
 
 When `principal(req)` resolves null, each client gets its own ephemeral session
-principal, carried by an opaque httpOnly cookie scoped to the wire base. The
-cookie is just a pointer; the session's `vendo_sessions` row is the authority.
-Sessions are per client, stored on disk like everything else, survive restarts,
-and expire through the idle-TTL sweep.
+principal, carried by an opaque httpOnly cookie set at `Path=/` — deliberately
+site-wide, not scoped to the wire base, so a host can mint the pointer on its
+own document response instead of re-minting a fresh subject on every page load.
+Over HTTPS the cookie takes the fixation-proof `__Host-` form, which requires
+`Path=/` anyway. The cookie is just a pointer; the session's `vendo_sessions`
+row is the authority. Sessions are per client, stored on disk like everything
+else, survive restarts, and expire through the idle-TTL sweep.
 
 The composition registers app capability tools in actions and guard-binds every
 tool registry passed to agent, apps, and automations.
@@ -104,7 +112,7 @@ export function vendoTools(
 export function vendoMastraTools(
   vendo: Vendo,
   options?: { include?: string[]; exclude?: string[] },
-): Promise<Record<string, MastraTool>>;
+): Promise<Record<string, VendoMastraTool>>;
 
 export const VENDO_PRINCIPAL_KEY = "vendo-principal"; // required per request; missing fails closed
 export const VENDO_SESSION_KEY = "vendo-session-id";  // optional audit continuity
